refactor: AuthorizationExtractor 추상화 및 AuthorizationType 패키지 이동
