+++ src/main/java/nextstep/auth/authentication/filter/chain/BearerTokenAuthenticationFilter.java
@@ -2,11 +2,10 @@ package nextstep.auth.authentication.filter.chain;
 
 import lombok.RequiredArgsConstructor;
 import nextstep.auth.authentication.AuthenticationToken;
-import nextstep.auth.authentication.AuthorizationExtractor;
-import nextstep.auth.authentication.AuthorizationType;
 import nextstep.auth.authentication.provider.AuthenticationProvider;
 import nextstep.auth.authentication.provider.ProviderManager;
 import nextstep.auth.authentication.provider.ProviderType;
+import nextstep.auth.authorization.extractor.AuthorizationExtractor;
 import nextstep.auth.context.Authentication;
 
 import javax.servlet.http.HttpServletRequest;
@@ -15,10 +14,11 @@ import javax.servlet.http.HttpServletRequest;
 public class BearerTokenAuthenticationFilter extends AuthenticationChainFilter {
 
     private final ProviderManager providerManager;
+    private final AuthorizationExtractor authorizationExtractor;
 
     @Override
     protected AuthenticationToken createToken(HttpServletRequest request) {
-        String authCredentials = AuthorizationExtractor.extract(request, AuthorizationType.BEARER);
+        String authCredentials = authorizationExtractor.extract(request);
         return new AuthenticationToken(authCredentials, authCredentials);
     }
