+++ src/main/java/nextstep/auth/authorization/extractor/BasicAuthorizationExtractor.java
@@ -1,0 +1,9 @@
+package nextstep.auth.authorization.extractor;
+
+public class BasicAuthorizationExtractor extends AuthorizationExtractor {
+
+    @Override
+    protected AuthorizationType findAuthorizationType() {
+        return AuthorizationType.BASIC;
+    }
+}
