+++ src/main/java/nextstep/auth/authorization/extractor/BearerAuthorizationExtractor.java
@@ -1,0 +1,9 @@
+package nextstep.auth.authorization.extractor;
+
+public class BearerAuthorizationExtractor extends AuthorizationExtractor {
+
+    @Override
+    protected AuthorizationType findAuthorizationType() {
+        return AuthorizationType.BEARER;
+    }
+}
